requirements.txt
aiohttp==3.8.5
openai==0.28.0
pandas==2.1.1
python-dotenv==1.0.0
scikit_learn==1.3.1
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
